JAMES-2341 Move DefaultMailboxes at API level
