+++ mailbox/api/src/main/java/org/apache/james/mailbox/DefaultMailboxes.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
-package org.apache.james.jmap;
+package org.apache.james.mailbox;
 
 import java.util.List;
